refactor AW-168: extra wide icon card relation to sections
