+++ nextjs/src/components/dynamic-zone/wrapper/sections/extra-wide-icon-cards-grid-section.tsx
