+++ strapi/src/api/page-partner-and-product/content-types/page-partner-and-product/schema.json
@@ -87,7 +87,8 @@
         "sections.two-column-section",
         "sections.video-section",
         "sections.video-with-text-section",
-        "sections.calendly-section"
+        "sections.calendly-section",
+        "sections.extra-wide-icon-cards-grid-section"
       ]
     }
   }
