+++ strapi/src/api/solutions-page/content-types/solutions-page/schema.json
@@ -87,7 +87,8 @@
         "sections.two-column-section",
         "sections.video-section",
         "sections.video-with-text-section",
-        "sections.calendly-section"
+        "sections.calendly-section",
+        "sections.extra-wide-icon-cards-grid-section"
       ]
     }
   }
